--- src/main/kotlin/com/growingtree/growingtreeserver/auth/model/response/SendCodeResponse.kt
@@ -1,5 +1,0 @@
-package com.growingtree.growingtreeserver.auth.model.response
-
-data class SendCodeResponse(
-    val code: String,
-)
