--- src/main/kotlin/com/growingtree/growingtreeserver/auth/model/response/SignInResponse.kt
@@ -1,5 +1,0 @@
-package com.growingtree.growingtreeserver.auth.model.response
-
-data class SignInResponse(
-    val authId: Long,
-)
