+++ src/main/kotlin/com/growingtree/growingtreeserver/dto/auth/request/SendCodeRequest.kt
@@ -1,0 +1,8 @@
+package com.growingtree.growingtreeserver.dto.auth.request
+
+import com.growingtree.growingtreeserver.enums.ValidateType
+
+data class SendCodeRequest(
+    val validateType: ValidateType,
+    val email: String,
+)
