+++ src/main/kotlin/com/growingtree/growingtreeserver/dto/auth/response/SendCodeResponse.kt
@@ -1,0 +1,5 @@
+package com.growingtree.growingtreeserver.dto.auth.response
+
+data class SendCodeResponse(
+    val code: String,
+)
