+++ src/main/kotlin/com/growingtree/growingtreeserver/dto/auth/response/SignInResponse.kt
@@ -1,0 +1,5 @@
+package com.growingtree.growingtreeserver.dto.auth.response
+
+data class SignInResponse(
+    val authId: Long,
+)
