+++ src/main/kotlin/com/growingtree/growingtreeserver/repository/MotionsRepository.kt
@@ -3,5 +3,4 @@ package com.growingtree.growingtreeserver.repository
 import com.growingtree.growingtreeserver.domain.Motions
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface MotionsRepository : JpaRepository<Motions, Long> {
-}
+interface MotionsRepository : JpaRepository<Motions, Long>
